tools/google-cloud-support-slackbot/requirements.txt
cryptography==43.0.1
firebase==3.0.1
firebase-admin==5.1.0
Flask==1.1.4
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.57.0
grpcio==1.53.2
requests==2.32.0
six==1.16.0
slackclient==2.9.3
slackeventsapi==2.2.1
Werkzeug==3.0.6
google.cloud.logging
gevent==23.9.0
